refactor: made code pretty and put test helper function in its own file
