+++ app/db/db_test.go
@@ -2,7 +2,6 @@ package db
 
 import (
 	"fmt"
-	"os"
 	"testing"
 	"time"
 )
@@ -19,40 +18,10 @@ func TestNewDB(t *testing.T) {
 	}
 }
 
-//TODO: write tests for ExecWithTimeout
-
-func verifySchemaTableExists(db *DB, targetTable string) error {
-	var count int
-	query := fmt.Sprintf("SELECT count(*) FROM sqlite_master WHERE type='table' AND name='%s'", targetTable)
-	err := db.Connection.QueryRow(query).Scan(&count)
-	if err != nil {
-		return fmt.Errorf("failed to query %s table: %v", targetTable, err)
-	}
-	if count != 1 {
-		return fmt.Errorf("expected 1 '%s' table, but found %d", targetTable, count)
-	}
-	return nil
-}
-
 func TestInitialize(t *testing.T) {
-	db, err := NewDB("file::memory:?cache=shared", 2*time.Second)
-	if err != nil {
-		t.Fatalf("failed to open DB: %v", err)
-	}
+	db := createDummyDB(t)
 
-	defer func() {
-		err := db.Close()
-		if err != nil {
-			t.Errorf("failed to close DB: %v", err)
-		}
-	}()
-
-	err = db.Initialize()
-	if err != nil {
-		t.Fatalf("failed to initialize DB: %v", err)
-	}
-
-	err = verifySchemaTableExists(db, "module")
+	err := verifySchemaTableExists(db, "module")
 	if err != nil {
 		t.Fatalf("failed to verify table existence: %v", err)
 	}
@@ -71,16 +40,7 @@ func TestInvalidNewDB(t *testing.T) {
 }
 
 func TestClose(t *testing.T) {
-	testFile := "test.db"
-
-	defer func() {
-		err := os.Remove(testFile)
-		if err != nil {
-			t.Fatalf("failed to remove test database file: %v", err)
-		}
-	}()
-
-	db, err := NewDB("file:"+testFile+"?cache=shared&_mode=rwc", 2*time.Second)
+	db, err := NewDB("file::memory:?cache=shared", 2*time.Second)
 	if err != nil {
 		t.Fatalf("failed to open DB: %v", err)
 	}
@@ -92,18 +52,9 @@ func TestClose(t *testing.T) {
 }
 
 func TestInitializeExistingTables(t *testing.T) {
-	db, err := NewDB("file::memory:?cache=shared", 2*time.Second)
-	if err != nil {
-		t.Fatalf("failed to open DB: %v", err)
-	}
-	defer db.Close()
+	db := createDummyDB(t)
 
-	err = db.Initialize()
-	if err != nil {
-		t.Fatalf("failed to initialize DB: %v", err)
-	}
-
-	err = db.Initialize()
+	err := db.Initialize()
 	if err != nil {
 		t.Fatalf("failed to initialize DB again: %v", err)
 	}
@@ -117,4 +68,17 @@ func TestInitializeExistingTables(t *testing.T) {
 	if err != nil {
 		t.Fatalf("failed to verify table existence: %v", err)
 	}
+}
+
+func verifySchemaTableExists(db *DB, targetTable string) error {
+	var count int
+	query := fmt.Sprintf("SELECT count(*) FROM sqlite_master WHERE type='table' AND name='%s'", targetTable)
+	err := db.Connection.QueryRow(query).Scan(&count)
+	if err != nil {
+		return fmt.Errorf("failed to query %s table: %v", targetTable, err)
+	}
+	if count != 1 {
+		return fmt.Errorf("expected 1 '%s' table, but found %d", targetTable, count)
+	}
+	return nil
 }
